requirementx.txt
Django==2.2.12
numpy==1.18.2
opencv-python==4.2.0.34
Pillow==7.1.1
